Ididistuff Update2
I added some variables, and found a different path to approach, and had
to overhaul my older script. This required me to go through my files and
figure out what I had been doing beforehand, and what my thought process
was going through with such things.
Of course, I figured it out, and equipped my script with the basics to
make it work, but a few rough edges need to be smoothed, things to be
implemented, and smoothed again. Soon I will be done.